requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
pycups==2.0.1
aiosqlite==0.19.0
pydantic[email]==2.5.0
jinja2==3.1.2
python-dotenv==1.0.0
